Organization and layout css
I added small directories in each section that I’ve work on. Each
section now has a number descriptor and the code its self has a
designation number to find items easier. This is a trial to see how I
like this form of organization. I have also finished the navigation,
added a profile section to adjust the profile picture and accompanying
paragraph and, started working on the main content or #portfolio.